Video Summarizer/requirements.txt
streamlit
agno
duckduckgo-search
python-dotenv
google-genai